[sparse] Factor BCOO-related routines into a separate submodule
